backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
transformers>=4.40.0
torch>=2.0.0
lime==0.2.0.1
numpy>=1.24.0
scikit-learn>=1.3.0
